add unit test for attaching and detaching new tab
