+++ streaming/src/main/scala/org/apache/spark/streaming/StreamingContext.scala
@@ -553,7 +553,6 @@ class StreamingContext private[streaming] (
    */
   def stop(stopSparkContext: Boolean = true): Unit = synchronized {
     stop(stopSparkContext, false)
-    this.uiTab.foreach(StreamingTab.detachStreamingTab(this, _))
   }
 
   /**
@@ -579,6 +578,7 @@ class StreamingContext private[streaming] (
     // Even if we have already stopped, we still need to attempt to stop the SparkContext because
     // a user might stop(stopSparkContext = false) and then call stop(stopSparkContext = true).
     if (stopSparkContext) sc.stop()
+    uiTab.foreach(StreamingTab.detachStreamingTab(this, _))
     // The state should always be Stopped after calling `stop()`, even if we haven't started yet:
     state = Stopped
   }
